reorganize entropy to match new dir structure (WIP)
This commit is a way point, the entropy concerns needs to be disentangled in the next commit
