+++ src/components/index.ts
@@ -1,10 +1,7 @@
 export { EntropicController } from "./entropic-controller"
-export { Entropy, useEntropyContext } from "../contexts/entropy"
+export { EntropyProvider as Entropy } from "../contexts"
 export { Field } from "./field"
 export { Matter } from "./matter"
-export {
-  PhysicsProvider as Physics,
-  usePhysics,
-} from "../contexts/physics-provider"
+export { PhysicsProvider as Physics } from "../contexts"
 export { Particle } from "./particle"
 export { Title } from "./title"
