+++ src/contexts/index.ts
@@ -1,3 +1,7 @@
 export { type Physics } from "./physics-context"
 export { PhysicsContext } from "./physics-context"
 export { PhysicsProvider } from "./physics-provider"
+
+export { type EntropyContextType } from "./entropy-context"
+export { EntropyContext } from "./entropy-context"
+export { EntropyProvider } from "./entropy-provider"
