+++ src/hooks/index.ts
@@ -1,3 +1,4 @@
 export { useEntropy } from "./use-entropy"
+export { useEntropyContext } from "./use-entropy-context"
 export { useInitialConditions } from "./use-initial-conditions"
 export { usePhysics } from "./use-physics"
